+++ web/jest.config.mjs
@@ -12,12 +12,6 @@ const clientTestConfig = {
   testEnvironment: "jest-environment-jsdom",
 };
 
-const unitTestConfig = {
-  displayName: "unit",
-  testMatch: ["/**/*.unit.test.[jt]s?(x)"],
-  testEnvironment: "jest-environment-jsdom",
-};
-
 const serverTestConfig = {
   displayName: "sync-server",
   testMatch: ["/**/*.servertest.[jt]s?(x)"],
@@ -55,12 +49,6 @@ const config = {
     await createJestConfig(clientTestConfig)(),
     {
       ...(await createJestConfig(serverTestConfig)()),
-      transformIgnorePatterns: [
-        `/web/node_modules/(?!(${esModules.join("|")})/)`,
-      ],
-    },
-    {
-      ...(await createJestConfig(unitTestConfig)()),
       transformIgnorePatterns: [
         `/web/node_modules/(?!(${esModules.join("|")})/)`,
       ],
